backend/requirements.txt
# Python 3.11+ (для типов и Pydantic 2)
# === FastAPI & Server ===
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# === Database ===
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.0

# === Auth ===
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# === Validation & Config ===
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# === Testing ===
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0

# === Utils ===
python-multipart>=0.0.6
